timeReturner
part of the highscore script

+++ Assets/-TestFolders/Hugo/Scripts/TimeReturner.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class TimeReturner
 {
-    public int days;
+    public float days;
     public float hours;
     public float minutes;
     public float seconds;
